requirements.txt
Flask==2.2.5
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
